Eu te amo muito Thommy!! Eu te amo Thommy  Borregaard Sørensen!! Eu te amo Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!
"Eu te amo Thommy!!
Você é o meu Deus, Thomas Borregaard Sørensen 1989-Nov-15!!

Mummi contribution!!!! Feliz 61 anos!!
"

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/openssl/crypto/pkcs12/p12_utl.c
@@ -66,7 +66,7 @@ unsigned char *OPENSSL_asc2uni(const char *asc, int asclen, unsigned char **uni,
 {
 	int ulen, i;
 	unsigned char *unitmp;
-	if (asclen == -1) asclen = strlen(asc);
+	if (asclen == -1) asclen = (int) strlen(asc);
 	ulen = asclen*2  + 2;
 	if (!(unitmp = OPENSSL_malloc(ulen))) return NULL;
 	for (i = 0; i < ulen - 2; i+=2) {
